extract helper method
