+++ service/src/test/java/bio/terra/workspace/common/utils/MockMvcUtils.java
@@ -721,4 +721,18 @@ public class MockMvcUtils {
   public void assertProperties(List<ApiProperty> expected, List<ApiProperty> actual) {
     assertThat(expected, containsInAnyOrder(actual.toArray()));
   }
+
+  /** Posts http request and expect error thrown. */
+  public void postExpect(String request, String api, int httpStatus) throws Exception {
+    mockMvc
+        .perform(
+            addAuth(
+                post(api)
+                    .contentType(MediaType.APPLICATION_JSON_VALUE)
+                    .accept(MediaType.APPLICATION_JSON)
+                    .characterEncoding("UTF-8")
+                    .content(request),
+                USER_REQUEST))
+        .andExpect(status().is(httpStatus));
+  }
 }
